refactor: レビュー指摘を反映しe2eヘルパー集約とCI権限・envノブを堅牢化
